requirements.txt
pinecone>=5,<6
flask
flask-cors
pillow
numpy
gunicorn
python-dotenv
torch
torchvision
firebase-admin
pyjwt
cryptography
PyJWT
